LPD-55581 - Rename to avoid confusion with /api submodule directories
