+++ modules/apps/asset/asset-list-web/src/main/resources/META-INF/resources/js/actions/index.js
